refactor EagerDataModel to abstract class, add RowExpansionDataModel
separate concerns and explicitly provide loadAll() for LazyDataModel
