+++ web/src/main/java/org/phoenixctms/ctsms/web/model/LazyDataModelBase.java
@@ -1,58 +1,28 @@
 package org.phoenixctms.ctsms.web.model;
 
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-import org.phoenixctms.ctsms.util.AssociationPath;
-import org.phoenixctms.ctsms.util.CommonUtil;
 import org.phoenixctms.ctsms.vo.PSFVO;
 import org.phoenixctms.ctsms.web.util.WebUtil;
-import org.primefaces.event.data.PageEvent;
-import org.primefaces.model.LazyDataModel;
 import org.primefaces.model.SelectableDataModel;
 import org.primefaces.model.SortOrder;
 
 public abstract class LazyDataModelBase<T> extends LazyDataModel<IDVO> implements SelectableDataModel<IDVO> {
 
 	private String currentPageIdsString;
-	private int page = 0;
 
 	public String getCurrentPageIds() {
 		return (currentPageIdsString != null ? currentPageIdsString : "");
 	}
 
-	protected abstract Collection<T> getLazyResult(PSFVO psf);
-
 	protected Long getPageId(IDVO idvo) {
 		return idvo.getId();
 	}
 
-	@Override
-	public void setRowIndex(int rowIndex) {
-		/*
-		 * the following is a workaround to avoid div by zero in case of pageSize=0
-		 * (possible bug in pf)
-		 * this.rowIndex = rowIndex == -1 ? rowIndex : (rowIndex % pageSize);
-		 */
-		if (rowIndex == -1 || getPageSize() == 0) {
-			super.setRowIndex(-1);
-		} else {
-			super.setRowIndex(rowIndex);
-		}
-	}
-
-	@Override
-	public IDVO getRowData() {
-		// http://primefaces.prime.com.tr/forum/viewtopic.php?f=3&t=11353
-		if (isRowAvailable()) {
-			return super.getRowData();
-		} else {
-			return null;
-		}
-	}
+	protected abstract Collection<T> getLazyResult(PSFVO psf);
 
 	@Override
 	public IDVO getRowData(String key) {
@@ -73,48 +43,10 @@ public abstract class LazyDataModelBase<T> extends LazyDataModel<IDVO> implement
 
 	@Override
 	public List<IDVO> load(int first, int pageSize, String sortField, SortOrder sortOrder, Map<String, String> filters) {
-		PSFVO psf = new PSFVO();
-		psf.setFirst(first);
-		psf.setPageSize(pageSize);
-		if (sortField != null && sortField.length() > 0) {
-			AssociationPath sortFieldAssociationPath = new AssociationPath(sortField);
-			sortFieldAssociationPath.dropFirstPathElement();
-			switch (sortOrder) {
-				case ASCENDING:
-					psf.setSortField(sortFieldAssociationPath.getFullQualifiedPropertyName());
-					psf.setSortOrder(true);
-					break;
-				case DESCENDING:
-					psf.setSortField(sortFieldAssociationPath.getFullQualifiedPropertyName());
-					psf.setSortOrder(false);
-					break;
-				case UNSORTED:
-					psf.setSortField(null);
-					break;
-				default:
-					psf.setSortField(null);
-					break;
-			}
-		} else {
-			psf.setSortField(null);
-		}
-		Map<String, String> preparedFilters;
-		if (filters != null && filters.size() > 0) {
-			preparedFilters = new HashMap<String, String>(filters.size());
-			Iterator<Map.Entry<String, String>> filterIt = filters.entrySet().iterator();
-			while (filterIt.hasNext()) {
-				Map.Entry<String, String> filter = filterIt.next();
-				AssociationPath filterFieldAssociationPath = new AssociationPath(filter.getKey());
-				filterFieldAssociationPath.dropFirstPathElement();
-				preparedFilters.put(filterFieldAssociationPath.getFullQualifiedPropertyName(), filter.getValue());
-			}
-		} else {
-			preparedFilters = new HashMap<String, String>();
-		}
-		psf.setFilters(preparedFilters);
-		return load(psf);
+		return load(createPSFVO(first, pageSize, sortField, sortOrder, filters, true));
 	}
 
+	@Override
 	protected List<IDVO> load(PSFVO psf) {
 		Collection<?> lazyResult = getLazyResult(psf);
 		IDVO.transformVoCollection(lazyResult);
@@ -151,41 +83,5 @@ public abstract class LazyDataModelBase<T> extends LazyDataModel<IDVO> implement
 			updateRowCount(psf);
 		}
 		return (List<IDVO>) lazyResult;
-	}
-
-	public List<IDVO> loadAll() {
-		return load(null);
-	}
-
-	public void updateRowCount() {
-		PSFVO psf = new PSFVO();
-		psf.setPageSize(0);
-		getLazyResult(psf);
-		updateRowCount(psf);
-	}
-
-	protected void updateRowCount(PSFVO psf) {
-		int rowCount = 0;
-		Long psfRowCount;
-		if (psf != null && (psfRowCount = psf.getRowCount()) != null) {
-			rowCount = CommonUtil.safeLongToInt(psfRowCount);
-		}
-		setRowCount(rowCount);
-	}
-
-	public int getPage() {
-		return page;
-	}
-
-	public void setPage(int page) {
-		this.page = page;
-	}
-
-	public boolean isStorePage() {
-		return true; //store page by default
-	}
-
-	public void handlePageChanged(PageEvent event) {
-		page = event.getPage();
 	}
 }
