+++ web/src/main/java/org/phoenixctms/ctsms/web/model/RowExpansionDataModel.java
@@ -1,0 +1,13 @@
+package org.phoenixctms.ctsms.web.model;
+import java.util.List;
+
+public interface RowExpansionDataModel {
+
+	public int getAllRowCount();
+
+	public List<IDVO> getAllRows();
+
+	public List<IDVO> loadAll();
+
+	public void resetRows();
+}
